InputConsumer_ iteration number 3?

+++ src/Camera.h
@@ -20,8 +20,8 @@ enum CameraMovement
   JUMP
 };
 
-const float32 PITCH = 0.0f;
-const float32 YAW = -90.0f;
+const float64 PITCH = 0.0f;
+const float64 YAW = -90.0f;
 const float32 CAMERA_SPEED = 2.5f;
 const float32 SENSITIVITY = 0.1f;
 const float32 ZOOM = 45.0f;
@@ -39,8 +39,8 @@ public:
   glm::vec3 WorldUp;
 
   // Eular Angles
-  float32 Yaw;
-  float32 Pitch;
+  float64 Yaw;
+  float64 Pitch;
 
   // Camera options
   float32 MovementSpeed;
@@ -56,8 +56,8 @@ public:
   // Constructor with vectors
   Camera(glm::vec3 position = glm::vec3(0.0f, 0.0f, 3.0f),
           glm::vec3 up = glm::vec3(0.0f, 1.0f, 0.0f),
-          float32 yaw = YAW,
-          float32 pitch = PITCH);
+          float64 yaw = YAW,
+          float64 pitch = PITCH);
 
   // Returns the view matrix calculated using Eular Angles and the LookAt Matrix
   glm::mat4 GetViewMatrix(float32 deltaTime);
@@ -66,7 +66,7 @@ public:
   void ProcessInput(CameraMovement direction);
   void ProcessLeftAnalog(int16 stickX, int16 stickY, GLboolean constrainPitch = true);
   // Processes input received from a mouse input system. Expects the offset value in both the x and y direction.
-  void ProcessMouseMovement(float32 xoffset, float32 yoffset, GLboolean constrainPitch = true);
+  void ProcessMouseMovement(float64 xoffset, float64 yoffset, GLboolean constrainPitch = true);
   void ProcessRightAnalog(int16 stickX, int16 stickY, GLboolean constrainPitch = true);
   // Processes input received from a mouse scroll-wheel event. Only requires input on the vertical wheel-axis
   void ProcessMouseScroll(float32 yoffset);
